refactor: reuse PartitionIter
Allows any code that interacts with a set of PartitionData to be
abstracted from the underlying data structure.
