+++ ingester2/src/partition_iter.rs
@@ -1,0 +1,20 @@
+use parking_lot::Mutex;
+use std::{fmt::Debug, sync::Arc};
+
+use crate::buffer_tree::partition::PartitionData;
+
+/// An abstraction over any type that can yield an iterator of (potentially
+/// empty) [`PartitionData`].
+pub trait PartitionIter: Send + Debug {
+    /// Return the set of partitions in `self`.
+    fn partition_iter(&self) -> Box<dyn Iterator<Item = Arc<Mutex<PartitionData>>> + Send>;
+}
+
+impl<T> PartitionIter for Arc<T>
+where
+    T: PartitionIter + Send + Sync,
+{
+    fn partition_iter(&self) -> Box<dyn Iterator<Item = Arc<Mutex<PartitionData>>> + Send> {
+        (**self).partition_iter()
+    }
+}
