+++ ingester2/src/wal/mod.rs
@@ -7,7 +7,3 @@
 pub(crate) mod rotate_task;
 mod traits;
 pub(crate) mod wal_sink;
-
-maybe_pub! {
-    pub use super::rotate_task::*;
-}
